HousePredictionPrices/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
joblib==1.1.0
lxml==4.6.4
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
spyne==2.13.16
sqlparse==0.4.2
suds-py3==1.4.5.0
threadpoolctl==3.0.0
